requirements.txt
alembic==1.14.1
altgraph==0.17.4
amqp==5.3.1
attrs==24.2.0
Authlib==1.4.0
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.18
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn~=22.0.0
h11==0.14.0
identify==2.6.7
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
limits==4.0.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
nodeenv==1.9.1
oauthlib==3.2.2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pefile==2023.2.7
platformdirs==4.3.6
pre_commit==4.1.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
selenium==4.26.1
setuptools==75.5.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
Telethon==1.38.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
